api_rest/requirements.txt
wheel
setuptools
psycopg2
flask
flask-jsonpify
flask-sqlalchemy
flask-restful
names